+++ rajawali/src/main/java/org/rajawali3d/primitives/Cylinder.java
@@ -78,7 +78,7 @@ public class Cylinder extends Object3D {
 
 		float[] vertices = new float[numVertices * 3];
 		float[] normals = new float[numVertices * 3];
-		int[] indices = new int[numIndices];
+		short[] indices = new short[numIndices];
 
 		int i, j;
 		int vertIndex = 0, index = 0;
@@ -103,12 +103,12 @@ public class Cylinder extends Object3D {
 					int b = (mSegmentsC + 1) * j + i - 1;
 					int c = (mSegmentsC + 1) * (j - 1) + i - 1;
 					int d = (mSegmentsC + 1) * (j - 1) + i;
-					indices[index++] = a;
-					indices[index++] = b;
-					indices[index++] = c;
-					indices[index++] = a;
-					indices[index++] = c;
-					indices[index++] = d;
+					indices[index++] = (short) a;
+					indices[index++] = (short) b;
+					indices[index++] = (short) c;
+					indices[index++] = (short) a;
+					indices[index++] = (short) c;
+					indices[index++] = (short) d;
 				}
 			}
 		}
