+++ rajawali/src/main/java/org/rajawali3d/primitives/Plane.java
@@ -203,7 +203,7 @@ public class Plane extends Object3D {
 		float[] colors = null;
 		if (mCreateVertexColorBuffer)
 			colors = new float[numVertices * 4];
-		int[] indices = new int[mSegmentsW * mSegmentsH * 6];
+		short[] indices = new short[mSegmentsW * mSegmentsH * 6];
 		int vertexCount = 0;
 		int texCoordCount = 0;
 
@@ -256,21 +256,21 @@ public class Plane extends Object3D {
 				int lr = ur + 1;
 
                 if (mUpAxis == Axis.X || mUpAxis == Axis.Z) {
-                    indices[indexCount++] = ur;
-                    indices[indexCount++] = lr;
-                    indices[indexCount++] = ul;
+                    indices[indexCount++] = (short) ur;
+                    indices[indexCount++] = (short) lr;
+                    indices[indexCount++] = (short) ul;
 
-                    indices[indexCount++] = lr;
-                    indices[indexCount++] = ll;
-                    indices[indexCount++] = ul;
+                    indices[indexCount++] = (short) lr;
+                    indices[indexCount++] = (short) ll;
+                    indices[indexCount++] = (short) ul;
                 } else {
-                    indices[indexCount++] = ur;
-                    indices[indexCount++] = ul;
-                    indices[indexCount++] = lr;
+                    indices[indexCount++] = (short) ur;
+                    indices[indexCount++] = (short) ul;
+                    indices[indexCount++] = (short) lr;
 
-                    indices[indexCount++] = lr;
-                    indices[indexCount++] = ul;
-                    indices[indexCount++] = ll;
+                    indices[indexCount++] = (short) lr;
+                    indices[indexCount++] = (short) ul;
+                    indices[indexCount++] = (short) ll;
                 }
 			}
 		}
